refactor(CU-#86bxy30n6): remove unused code
