+++ src/components/DatePicker/index.ts
@@ -3,4 +3,5 @@ export {
     DatePickerFormContainer,
     DatePickerStyled,
     TodoToltipContainer,
+    TodoIconStyled,
 } from "@/components/DatePicker/DatePicker.styled";
